initial work on converting some of the values from there long long form to a usable currency

+++ src/cpp/services/tradeOperations.h
@@ -8,7 +8,7 @@
 struct liveShares {
     int quantity;
     long long price;
-    double profit;
+    long long profit;
     long long cost;
     long long priceBrought;
